Changed the testing file to an usual one.
There were two problems.

1. The previous version used /dev/random which is a device file. It is
   readable at any time, but its size is always zero.
2. Java does not have any functions to poll(2), so poll(2) in the slave
   is implemented with comparing file position and size.

As a result of 1. and 2., Java cannot poll(2) for /dev/random. I could
not fix this, but the main usage of Java version of fsyscall is on
Android. In Android, opening a device file on a slave is rare, I guess.
So I decided to avoid this problem without fixing.

Now, the test uses the usual file to read.