Changed not to run sigaction(2) in slaves.
Because the fslave sends all signals to the master.

This patch also encloses all system calls with suspend_signal() and
resume_signal() to apply the signal mask. The fslave cannot apply the
mask at sigprocmask(2) call. Because the fslave must send all signals
to the master even in waiting any system call requests in mainloop().