Changed to open /etc/pwd.db in the slave side.
Because the slave responses getuid(2), getgid(2) and so on, the
/etc/pwd.db in the slave must be used.

java/src/main/resources/.../pwd.db must be in the repository. Because
the mechanism of pwd.db and master.passwd is for *BSD. Linux is using
a different system, and it does not have pwd_mkdb(8). In development for
Android, Linux cannot generate pwd.db.