Changed to react for death by signal.
If a slave died by a signal such as SIGKILL, the slave could not send
any SIGNALEDs. This patch changes fshub to detect death of signaled
slaves with read(2) error. When a slave died without exit(2) nor
thr_exit(2), the fshub assumes that the slave was kill(2)'ed, and the
fshub send a SIGNALED.

This patch also changes fmhub with the same way. This way has one merit;
fmhub does not need to know how masters died. When a master exit(2)'ed,
all the fmhub must do is closing the pipe for read(2) error. When a
master was kill(2)'ed, all the fmhub must do is same. If the fmhub does
not use this way, the fmhub would have to know which signal kills the
master.