Fixed the fcntl(2) not to set the FD_CLOEXEC on a local file.
A local file can be shared by two or more virtual files. If the fcntl(2)
sets the FD_CLOEXEC on a local file, the local file will be closed even
if that file is used by some virtual files.

Now the FD_CLOEXEC flag for master file is held in fmaster. The local
file will be closed after execve(2) if no virtual files are using it.