Libraried fmhub.
I had guessed that fmhub can use an SSL socket in the same way of plain
sockets.  I could give an SSL socket to an fmhub after execve(2), the
fmhub could use the same code to {read,write} that socket, I had guessed
so.  In my old plan, after nexecd handshakes with a client over SSL,
the nexecd passes the connection to an fmhub process.  The fmhub would
not need to pay attention on kind of the connection.

However, it is impossible.  OpenSSL stores something in memory to
{encrypt,decrypt} for performance.  An execve(2) call will take these
data away.  So fmhub processes must work in the same process which did
SSL handshake.

The new library, libfmhub will be used by nexecd.  After fork(2) and SSL
handshake, nexecd processes call fmhub_run().