Fixed to shutdown fmhub for unexpected socket closing.
If a slave closes the connection without SSL_shutdown(3) call,
io_select() call in the fmhub never stop. Because select(2) says that
the socket is ready to read (a read(2) call will return zero for EOF),
however, SSL_pending(3) returns zero. So the for loop in io_select()
never end. This patch enhances error detection in is_readable_ssl() to
fix this bug.